+++ docs/usage/ReactGlobe.tsx
@@ -1,0 +1,11 @@
+import * as React from 'react';
+
+import ReactGlobeSrc from '../../src';
+
+export default function ReactGlobe({ ...otherProps }): React.ReactNode {
+  return (
+    <div style={{ height: 400, width: '100%' }}>
+      <ReactGlobeSrc {...otherProps} />
+    </div>
+  );
+}