+++ src/hooks/useCamera.ts
@@ -4,10 +4,21 @@ import * as THREE from 'three';
 import OrbitControls from 'three-orbitcontrols';
 
 import {
-    CAMERA_DAMPING_FACTOR, CAMERA_FAR, CAMERA_FOV, CAMERA_MIN_DISTANCE_RADIUS_SCALE, CAMERA_NEAR,
-    RADIUS
+  CAMERA_DAMPING_FACTOR,
+  CAMERA_FAR,
+  CAMERA_FOV,
+  CAMERA_MIN_DISTANCE_RADIUS_SCALE,
+  CAMERA_NEAR,
+  RADIUS,
 } from '../defaults';
-import { CameraOptions, Coordinates, FocusOptions, LightOptions, Position, Size } from '../types';
+import {
+  CameraOptions,
+  Coordinates,
+  FocusOptions,
+  LightOptions,
+  Position,
+  Size,
+} from '../types';
 import { coordinatesToPosition, tween } from '../utils';
 
 const { useEffect, useRef } = React;