+++ src/index.ts
@@ -1,2 +1,10 @@
-export * from './defaults';
+export {
+  defaultBarMarkerOptions,
+  defaultDotMarkerOptions,
+  defaultCameraOptions,
+  defaultFocusOptions,
+  defaultGlobeOptions,
+  defaultLightOptions,
+  defaultMarkerOptions,
+} from './defaults';
 export { default } from './ReactGlobe';